pyproject.toml
[tool.poetry]
name = "re-bingo"
version = "0.1.0"
description = ""
authors = ["Muspi Merol <admin@muspimerol.site>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.89.1"
orjson = "^3.8.5"
strawberry-graphql = "^0.154.0"
tortoise-orm = "^0.19.2"
uvicorn = "^0.20.0"
aiomysql = "^0.1.1"
loguru = "^0.6.0"
pyyaml = "^6.0"
watchfiles = "^0.18.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }

[tool.poetry.group.dev.dependencies]
aerich = "^0.7.1"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "src.db.SQL_CONFIG"
location = "./migrations"
src_folder = "."
